Add take picture and make storage folder functions
Add a small check_image_folder() function that checks if the storage
folder exists.

Add a take_pics() function that initializes the camera, takes the
specified number of pictures, and stores them in the storage folder.
Put the camera functionality and take_pics() function into the
main_screen() function.

Import the datetime, pathlib and picamera modules

Create three settings variables: path to simplebooth app, path to picture
storage folder, and number of pictures to take.

Add settings for the camera.